fix: Use the .toFloat() instead

+++ app/src/main/java/com/baidaidai/rootless_store/components/stratScreen/components/HosterStatusProgressRow.kt
@@ -41,7 +41,7 @@ fun HosterStatusProgressRow(label: String, currentValue: Double,maxValue: Double
         ) {
             LinearProgressIndicator(
                 progress = {
-                    currentValueProgress
+                    currentValueProgress.toFloat()
                 },
                 drawStopIndicator = {}
             )
